+++ benchmarks/definition/__init__.py
@@ -1,0 +1,13 @@
+"""Benchmark task definitions."""
+
+from benchmarks.definition.config import (
+    Benchmark,
+    BenchmarkSettings,
+    ConvergenceExperimentSettings,
+)
+
+__all__ = [
+    "ConvergenceExperimentSettings",
+    "Benchmark",
+    "BenchmarkSettings",
+]