+++ examples/Basics/recommenders.py
@@ -79,7 +79,7 @@ available_acq_functions = [
     "UCB",  # upper confidence bound with beta of 1.0
 ]
 
-# Note that the qvailability of the acquisition functions might depend on the `batch_size`:
+# Note that the availability of the acquisition functions might depend on the `batch_size`:
 #   - If `batch_size` is set to 1, all available acquisition functions can be chosen
 #   - If a larger value is chosen, only those that allow batching.
 #       That is, 'q'-variants of the acquisition functions must be chosen.